[Refactor] 컴포넌트 결합
